third_party/blink/web_tests/wpt_internal/attribution-reporting/event-level-debug-report.sub.https.html

<!doctype html>
<meta charset=utf-8>
<meta name=timeout content=long>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/helpers.js"></script>
<script>
attribution_reporting_promise_test(async t => {
  const source_debug_key = '246';
  const trigger_debug_key = '357';
  await registerAttributionSrc(t, {
    source: {
      destination: `https://{{host}}`,
      debug_key: source_debug_key,
    },
    cookie: attributionDebugCookie,
  });
  await registerAttributionSrc(t, {
    trigger: {
      event_trigger_data: [{}],
      debug_key: trigger_debug_key,
    },
    cookie: attributionDebugCookie,
  });

  const payload = await pollEventLevelReports();
  assert_equals(payload.reports.length, 1);
  const report = JSON.parse(payload.reports[0].body);
  assert_equals(report.source_debug_key, source_debug_key);
  assert_equals(report.trigger_debug_key, trigger_debug_key);

  const debugPayload = await pollEventLevelDebugReports();
  assert_equals(debugPayload.reports.length, 1);
  const debugReport = JSON.parse(debugPayload.reports[0].body);
  assert_equals(debugReport.source_debug_key, source_debug_key);
  assert_equals(debugReport.trigger_debug_key, trigger_debug_key);
}, 'Debug report is received if source and trigger debug keys are set.');
</script>
